Refactored in the new Job object and input output file tests for restart

+++ lib/Breakdancer.pm
@@ -48,20 +48,22 @@ sub bam2cfg {
       $stdDevCutoff = 3;
     }
   
-    my $outDate = -M $output;
-    my $inDate = -M $sampleBAM;
-  
-    my $command;
-    # -M gives modified date relative to now. The bigger the older.
-    #if(!defined($outDate) || !defined($inDate) || $inDate < $outDate) {
-        $command .= 'module load '.LoadConfig::getParam($rH_cfg, 'bam2cfg', 'moduleVersion.breakdancer').' ;';
-        $command .= ' module load '.LoadConfig::getParam($rH_cfg, 'bam2cfg', 'moduleVersion.samtools').' ;';
-        $command .= ' bam2cfg.pl -g -h ';
-        $command .= ' -c '.$stdDevCutoff;
-        $command .= ' '.$sampleBAM;
-        $command .= ' > '.$output;
-    #}
-    return $command;
+    my $up2date = PipelineUtils::testInputOutputs([$sampleBAM], [$output]);
+    my $ro_job = new Job(!defined($up2date));
+
+    if (!$ro_job->isUp2Date()) {
+      my $command;
+      $command .= 'module load '.LoadConfig::getParam($rH_cfg, 'bam2cfg', 'moduleVersion.breakdancer').' ;';
+      $command .= ' module load '.LoadConfig::getParam($rH_cfg, 'bam2cfg', 'moduleVersion.samtools').' ;';
+      $command .= ' bam2cfg.pl -g -h ';
+      $command .= ' -c '.$stdDevCutoff;
+      $command .= ' '.$sampleBAM;
+      $command .= ' > '.$output;
+      $command .= ' ' . $up2date;
+
+      $ro_job->addCommand($command);
+    }
+    return $ro_job;
 }
 
 sub pairedBRDITX {
@@ -70,20 +72,22 @@ sub pairedBRDITX {
     my $inputCFG        = shift;
     my $outputPrefix    = shift;
 
-    my $outDate = -M $outputPrefix.'.ctx';
-    my $inDate = -M $inputCFG;
-  
-    my $command;
-    # -M gives modified date relative to now. The bigger the older.
-    #if(!defined($outDate) || !defined($inDate) || $inDate < $outDate) {
-        $command .= 'module load '.LoadConfig::getParam($rH_cfg, 'pairedBRDITX', 'moduleVersion.breakdancer').' ;';
-        $command .= ' breakdancer_max '.LoadConfig::getParam($rH_cfg, 'pairedBRDITX', 'brdParameters');
-        $command .= ' -g '.$outputPrefix.'.bed';
-        $command .= ' -d '.$outputPrefix.'.ctx';
-        $command .= ' '.$inputCFG;
-        $command .= ' > '.$outputPrefix.'.ctx';
-    #}
-    return $command;
+    my $up2date = PipelineUtils::testInputOutputs([$inputCFG], [$outputPrefix.'.bed', $outputPrefix.'.ctx']);
+    my $ro_job = new Job(!defined($up2date));
+
+    if (!$ro_job->isUp2Date()) {
+      my $command;
+      $command .= 'module load '.LoadConfig::getParam($rH_cfg, 'pairedBRDITX', 'moduleVersion.breakdancer').' ;';
+      $command .= ' breakdancer_max '.LoadConfig::getParam($rH_cfg, 'pairedBRDITX', 'brdParameters');
+      $command .= ' -g '.$outputPrefix.'.bed';
+      $command .= ' -d '.$outputPrefix.'.ctx';
+      $command .= ' '.$inputCFG;
+      $command .= ' > '.$outputPrefix.'.ctx';
+      $command .= ' ' . $up2date;
+
+      $ro_job->addCommand($command);
+    }
+    return $ro_job;
 }
 
 sub pairedBRD {
@@ -93,36 +97,42 @@ sub pairedBRD {
     my $inputCFG        = shift;
     my $outputPrefix    = shift;
 
-    my $outDate = -M $outputPrefix.'.ctx';
-    my $inDate = -M $inputCFG;
-  
-    my $command;
-    # -M gives modified date relative to now. The bigger the older.
-    #if(!defined($outDate) || !defined($inDate) || $inDate < $outDate) {
-    $command .= 'module load '.LoadConfig::getParam($rH_cfg, 'pairedBRD', 'moduleVersion.breakdancer').' ;';
-    $command .= ' breakdancer_max '.LoadConfig::getParam($rH_cfg, 'pairedBRD', 'brdParameters');
-    $command .= ' -o '.$chr;
-    $command .= ' -g '.$outputPrefix.'.bed';
-    $command .= ' -d '.$outputPrefix.'.ctx';
-    $command .= ' '.$inputCFG;
-    $command .= ' > '.$outputPrefix.'.ctx';
-    #}
-    return $command;
+    my $up2date = PipelineUtils::testInputOutputs([$inputCFG], [$outputPrefix.'.bed', $outputPrefix.'.ctx']);
+    my $ro_job = new Job(!defined($up2date));
+
+    if (!$ro_job->isUp2Date()) {
+      my $command;
+      $command .= 'module load '.LoadConfig::getParam($rH_cfg, 'pairedBRD', 'moduleVersion.breakdancer').' ;';
+      $command .= ' breakdancer_max '.LoadConfig::getParam($rH_cfg, 'pairedBRD', 'brdParameters');
+      $command .= ' -o '.$chr;
+      $command .= ' -g '.$outputPrefix.'.bed';
+      $command .= ' -d '.$outputPrefix.'.ctx';
+      $command .= ' '.$inputCFG;
+      $command .= ' > '.$outputPrefix.'.ctx';
+      $command .= ' ' . $up2date;
+
+      $ro_job->addCommand($command);
+    }
+    return $ro_job;
 }
 
 sub mergeCTX {
     my $rH_cfg          = shift;
     my $outputPrefix    = shift;
 
-    my $command;
-    # -M gives modified date relative to now. The bigger the older.
-    #if(!defined($outDate) || !defined($inDate) || $inDate < $outDate) {
-    $command .= 'rm ' .$outputPrefix .'.ctx && ' ;
-    $command .= 'touch ' .$outputPrefix .'.ctx && ' ;
-    $command .= 'for i in ' .$outputPrefix .'.*.ctx ; do cat \$i >> '  .$outputPrefix .'.ctx' ;
-    #}
-    return $command;
-}
+    my $up2date = PipelineUtils::testInputOutputs(undef, [$outputPrefix .'.ctx']);
+    my $ro_job = new Job(!defined($up2date));
 
-    
+    if (!$ro_job->isUp2Date()) {
+      my $command;
+      $command .= 'rm ' .$outputPrefix .'.ctx && ' ;
+      $command .= 'touch ' .$outputPrefix .'.ctx && ' ;
+      $command .= 'for i in ' .$outputPrefix .'.*.ctx ; do cat \$i >> '  .$outputPrefix .'.ctx' ;
+      $command .= ' ' . $up2date;
+
+      $ro_job->addCommand($command);
+    }
+    return $ro_job;
+}
+ 
 1;
