+++ lib/Metrics.pm
@@ -49,12 +49,13 @@ sub rnaQc{
   my $outputFolder     = shift;
 
 
-  my $latestFile = -M $inputFile;
   my $outputIndexFile= $outputFolder. 'index.html';
 
-  my $command;
-  # -M gives modified date relative to now. The bigger the older.
-  if(!defined($latestFile) || !defined(-M $outputIndexFile) || $latestFile < -M $outputIndexFile) {
+  my $up2date = PipelineUtils::testInputOutputs([$inputFile], [$outputFolder .'.zip', $outputIndexFile]);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
+    my $command;
     $command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'rnaQc','moduleVersion.bwa') ;
     $command .= ' ' .LoadConfig::getParam($rH_cfg, 'rnaQc','moduleVersion.rnaseqc') .' &&';
     $command .= ' java -Djava.io.tmpdir='.LoadConfig::getParam($rH_cfg, 'rnaQc', 'tmpDir').' '.LoadConfig::getParam($rH_cfg, 'rnaQc', 'extraJavaFlags').' -Xmx'.LoadConfig::getParam($rH_cfg, 'rnaQc', 'metricsRam').' -jar \${RNASEQC_JAR}';
@@ -64,10 +65,13 @@ sub rnaQc{
     $command .= ' -r ' .LoadConfig::getParam($rH_cfg, 'rnaQc','referenceFasta');
     $command .= ' -o ' .$outputFolder ;
     $command .= ' -BWArRNA ' .LoadConfig::getParam($rH_cfg, 'rnaQc','ribosomalFasta') .' &&';
-    $command .= ' zip -r ' .$outputFolder .'.zip ' .$outputFolder
+    $command .= ' zip -r ' .$outputFolder .'.zip ' .$outputFolder;
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
   }
     
-  return $command;
+  return $ro_job;
 }
 
 sub saturation {
@@ -78,19 +82,21 @@ sub saturation {
 	my $saturationDir = shift;
 
 
-	my $latestFile = -M $countFile;
-	my $testfile = -M $saturationDir .'.zip ';
-	
-	
-	my $command;
-	if(!defined($latestFile) || !defined($testfile) || $latestFile < $testfile) {
+	my $up2date = PipelineUtils::testInputOutputs([$countFile], [$saturationDir .'.zip']);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
+  	my $command;
 		$command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'saturation' , 'moduleVersion.cranR') .' ' . LoadConfig::getParam($rH_cfg, 'saturation' , 'moduleVersion.tools') . ' &&';
 		$command .= ' Rscript \$R_TOOLS/rpkmSaturation.R ' .$countFile .' ' .$geneSizeFile .' ' .$rpkmDir .' ' .$saturationDir;
 		$command .= ' ' .LoadConfig::getParam($rH_cfg, 'saturation' , 'threadNum');
 		$command .= ' ' .LoadConfig::getParam($rH_cfg, 'saturation' , 'optionR').' &&';
-		$command .= ' zip -r ' .$saturationDir .'.zip ' .$saturationDir
+		$command .= ' zip -r ' .$saturationDir .'.zip ' .$saturationDir;
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
 	}
-	return $command;
+	return $ro_job;
 }
 
 sub fpkmCor {
@@ -99,11 +105,18 @@ sub fpkmCor {
 	my $folderFile     = shift;
 	my $outputBaseName = shift;
 	
-	my $command;
-	$command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.cranR') .' ' . LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.tools') . ' ;';
-	$command .= ' Rscript \$R_TOOLS/fpkmStats.R ' .$paternFile .' ' .$folderFile .' ' .$outputBaseName;
-	
-	return $command;
+  my $up2date = PipelineUtils::testInputOutputs([$paternFile], [$outputBaseName]);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
+  	my $command;
+	  $command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.cranR') .' ' . LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.tools') . ' ;';
+  	$command .= ' Rscript \$R_TOOLS/fpkmStats.R ' .$paternFile .' ' .$folderFile .' ' .$outputBaseName;
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
+	}
+	return $ro_job;
 }
 
 sub readStats {
@@ -113,11 +126,11 @@ sub readStats {
 	my $sampleName = shift;
 	my $fileType = shift;
 	
-	my $latestInputFile = -M $inputFile;
-	my $latestOutputFile = -M $outputFile;
-	
-	my $command;
-	if(!defined($latestInputFile) || !defined($latestOutputFile) || $latestInputFile <  $latestOutputFile) {
+  my $up2date = PipelineUtils::testInputOutputs([$inputFile], [$outputFile]);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
+  	my $command;
 		if ((lc $fileType) eq "trim") {
 			$command .= 'grep \"Input Read\" ' .$inputFile;
 			$command .= ' | awk -F\":\" -v na='.$sampleName .' \' BEGIN { OFS=\"\t\" } { split(\$2,a,\" \"); split(\$3,b,\" \"); print na,a[1],b[1]} \''; 
@@ -134,9 +147,12 @@ sub readStats {
 			$command .= ' | wc -l  > ' .$outputFile;
 			#----
 		}
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
 	}
 
-	return $command;
+	return $ro_job;
 }
 
 
@@ -149,20 +165,23 @@ sub mergeIndvidualReadStats{
 	my $alignFile = shift;
 	my $outputFile =shift;
 	
-	my $latestInputFile = -M $alignFile;
-	my $latestOutputFile = -M $outputFile;
-	
-	my $command;
-	if(!defined($latestInputFile) || !defined($latestOutputFile) || $latestInputFile <  $latestOutputFile) {
+  my $up2date = PipelineUtils::testInputOutputs([$alignFile], [$outputFile]);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
+	  my $command;
 		$command .= 'echo \"' .$sampleName .'\"' ;
 		$command .= ' | cat - ' .$rawFile .' ' .$filterFile .' ' .$alignFile;
 		$command .= ' | tr \'\n\' \',\' ';
 		$command .= ' | awk \' BEGIN {FS=\",\"} { print \$1 \",\" \$2 \",\" \$3 \",\" \$4}\''; 
 		$command .= ' > ' .$outputFile .' &&'; 
 		$command .= ' rm  ' .$rawFile .' ' .$filterFile .' ' .$alignFile;
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
 	}
 
-	return $command;
+	return $ro_job;
 }
 
 sub mergeReadStats{
@@ -170,13 +189,20 @@ sub mergeReadStats{
 	my $paternFile     = shift;
 	my $folderFile     = shift;
 	my $outputFile     = shift;
-	
+
+  my $up2date = PipelineUtils::testInputOutputs([$paternFile], [$outputFile]);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
 	my $command;
-	$command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.cranR') .' ' . LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.tools') . ' ;';
-	$command .= ' Rscript \$R_TOOLS/mergeReadStat.R ' .$paternFile .' ' .$folderFile .' ' .$outputFile .' &&';
-	$command .= ' rm ' .$folderFile .'/*' .$paternFile;
-	
-	return $command;
+  	$command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.cranR') .' ' . LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.tools') . ' ;';
+	  $command .= ' Rscript \$R_TOOLS/mergeReadStat.R ' .$paternFile .' ' .$folderFile .' ' .$outputFile .' &&';
+  	$command .= ' rm ' .$folderFile .'/*' .$paternFile;
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
+	}
+	return $ro_job;
 }
 
 
@@ -190,11 +216,20 @@ sub mergeTrimmomaticStats{
 	if (!defined($libraryType) || $libraryType eq "") {
 		$libraryType= 'unknown';
 	}
-	my $command;
-	$command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.cranR') .' ' . LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.tools') . ' ;';
-	$command .= ' Rscript \$R_TOOLS/mergeTrimmomaticStat.R ' .$paternFile .' ' .$folderFile .' ' .$outputFile .' ' .$libraryType;
+
+  my $up2date = PipelineUtils::testInputOutputs([$paternFile], [$outputFile]);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
+  	my $command;
+	  $command .= 'module load ' .LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.cranR') .' ' . LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.tools') . ' ;';
+  	$command .= ' Rscript \$R_TOOLS/mergeTrimmomaticStat.R ' .$paternFile .' ' .$folderFile .' ' .$outputFile .' ' .$libraryType;
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
+  }
 	
-	return $command;
+	return $ro_job;
 }
 
 sub mergePrintReadStats{
@@ -205,17 +240,20 @@ sub mergePrintReadStats{
 	my $outputFile  				= shift;
 	my $command							= undef;
 
-	my $latestInputFile = -M $flagStatsFile;
-	my $latestOutputFile = -M $outputFile;
-	
-	if(!defined($latestInputFile) || !defined($latestOutputFile) || $latestInputFile <  $latestOutputFile) {
+  my $up2date = PipelineUtils::testInputOutputs([$flagStatsFile], [$outputFile]);
+  my $ro_job = new Job(!defined($up2date));
+
+  if (!$ro_job->isUp2Date()) {
 		$command .= 'module load ' . LoadConfig::getParam($rH_cfg, 'metrics' , 'moduleVersion.tools') . ' ;';
 		$command .= ' perl -MReadMetrics -e \' ReadMetrics::mergeStats(\"'.$sampleName.'\",';
 		$command .= ' \"'. $outputFile .'\", ReadMetrics::parseTrimOutput(\"'.$sampleName.'\",';
 		$command .= ' \"'. $trimOutputFile .'\"), ReadMetrics::parseFlagstats(\"'.$sampleName.'\",';
 		$command .= ' \"'. $flagStatsFile .'\"));\'';
+    $command .= ' ' . $up2date;
+
+    $ro_job->addCommand($command);
 	}
-	return $command;
+	return $ro_job;
 }
 
 1;
